Remove the references to Python's Scala codes in R's Scala codes
